Adding DOC string and Optimazing tokenize function
I added a DOC string for each function in the train_classifier.py program. Also, I optimazied the search for stopwords in the tokenize function. Changed the classifier to test a simpler but more tuned one. Added a time counter just to keep track of how slow is to train the model.

Regarding the tokenize function, It was too time consuming after the stopword removal inclusion. Added a hash search by a stopword dict so It would get a bit faster. Also, changede the clean token process before the stopword check.