Included new feature and changed the toknize function
This version included a removal named entities and url replacer function to the tokenize function. Also added a categorical feature with the genre of each message.

Before this change, some named entities were flooding the features vector with separeted tokens that wouldn't say much about each message. Used parts of speach and named entities functions of nltk to this. Also, It was possible to remove some parts of speach like cardinal numbers and etc.

The inclusion of the genre category in the set of features aimed to separate evaluation of the messages according of who is speaking in the message. Genaraly a news message speaks in a different way about the same issue a person would do. So the words weights for each type os message should differ.