Changes in Classifier and GridSearch
I changed the classifier to AdaBoost in order to solve the imbalanced label data *. Also added a selectpercentile step in the pipeline.

The parameters of GridSearch are commetend because it was too time consuming running the grid. So I left in the defaut set which is the choosed one after the Grid.

*(https://www.analyticsvidhya.com/blog/2017/03/imbalanced-data-classification/#:~:text=Dealing%20with%20imbalanced%20datasets%20entails,as%20it%20has%20wider%20application.)